@@ -18,8 +18,8 @@ export const routes: Routes = [
     path: '', component: LayoutComponent, data: { title: '' },
     children: [
       { path: 'dashboard', component: DashboardComponent, data: { title: 'Dashboard' } },
-      { path: 'product', loadChildren: './views/product/product.module#ProductModule' },
-      { path: 'category', loadChildren: './views/category/category.module#CategoryModule' }
+      { path: 'product', loadChildren: () => import('./views/product/product.module').then(m => m.ProductModule) },
+      { path: 'category', loadChildren: () => import('./views/category/category.module').then(m => m.CategoryModule) }
     ],
     canActivateChild: [AuthGuardService],
     canActivate: [AuthGuardService]
